move MasterSlaveRuleConfiguration to master-slave-core-api module
