+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/rule/MasterSlaveDataSourceRule.java
@@ -19,8 +19,8 @@ package org.apache.shardingsphere.core.rule;
 
 import lombok.AccessLevel;
 import lombok.Getter;
-import org.apache.shardingsphere.api.config.masterslave.LoadBalanceStrategyConfiguration;
-import org.apache.shardingsphere.api.config.masterslave.MasterSlaveDataSourceConfiguration;
+import org.apache.shardingsphere.masterslave.api.config.LoadBalanceStrategyConfiguration;
+import org.apache.shardingsphere.masterslave.api.config.MasterSlaveDataSourceConfiguration;
 import org.apache.shardingsphere.spi.ShardingSphereServiceLoader;
 import org.apache.shardingsphere.spi.masterslave.MasterSlaveLoadBalanceAlgorithm;
 import org.apache.shardingsphere.spi.type.TypedSPIRegistry;
