+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/yaml/swapper/root/RuleRootConfigurationsYamlSwapper.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.core.yaml.swapper.root;
 
-import org.apache.shardingsphere.api.config.masterslave.MasterSlaveRuleConfiguration;
+import org.apache.shardingsphere.masterslave.api.config.MasterSlaveRuleConfiguration;
 import org.apache.shardingsphere.api.config.shadow.ShadowRuleConfiguration;
 import org.apache.shardingsphere.api.config.sharding.ShardingRuleConfiguration;
 import org.apache.shardingsphere.core.yaml.config.YamlRootRuleConfigurations;
